% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_metaanalysis.R
\name{create_dataset}
\alias{create_dataset}
\title{Generating the dataset for the meta-analysis}
\usage{
create_dataset(
  ...,
  col_names = "GeneName",
  col_adjpvals = c("adj.P.Val", "padj"),
  col_log2FC = c("logFC", "log2FoldChange")
)
}
\arguments{
\item{...}{dataframes with the differential expression analyses to be included in the final dataset.}

\item{col_names}{character vector with all the possible column names that indicate the gene names in the input dataframes. If this column is absent the rownames will be used as gene names.}

\item{col_adjpvals}{character vector with all the possible column names that indicate the adjusted p-value in the input dataframes.}

\item{col_log2FC}{character vector with all the possible column names that indicate the log2 fold changes in the input dataframes.}
}
\value{
dataset list of lists, including a list called \code{names} containing sublists, each corresponding to a specific comparison from a study, of differentially expressed genes (each sublist is a character vector). The sublists are named with the names given to each of the objects given as differential expression analyses.
}
\description{
Given a number of differential expression analyses it creates a dataset with the right format for the downstream analyses.
}
\examples{
# import dataset with DESeq2 differential expression analysis
data("ipsc_deseq2")
# import dataset with limma differential expression analysis
data("thymus_limma")
list_array_2studies=create_dataset(thymus_limma, ipsc_deseq2,col_names="GeneName", col_adjpvals=c("adj.P.Val", "padj"), col_log2FC=c("logFC", "log2FoldChange"))
}
