% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_metaanalysis.R
\name{find_coDE}
\alias{find_coDE}
\title{Find co-differentially expressed genes (co-DE genes)}
\usage{
find_coDE(gene, dataset)
}
\arguments{
\item{gene}{a gene (character) included in one of the sublists in dataset. Can also be a vector of more than one gene, in that case only co-DE events including the genes in the vector will be reported.}

\item{dataset}{list of lists, including a list called \code{names} containing sublists, each corresponding to a specific comparison from a study, of differentially expressed genes (each sublist is a character vector). The sublists should be named to identify the comparisons.}
}
\value{
a dataframe with 4 columns: \code{gene1} and \code{gene2} represent a co-DE gene pair; \code{pair_occurrency} indicates in how many comparisons a given co-DE gene pair is reported; \code{comparisons} lists these comparisons separated by ";". Results are sorted per \code{pair_occurrency}.
}
\description{
Find all genes that are found differentially expressed in the same comparison(s).
}
\examples{
data(list_array) #load data
# create a list of lists were only the first 500 most significant genes with adjusted p-value < 0.05 and fold change >1.5 or < -1.5 are included
data(list_array) #load data
list_array.05_fc1.5_max500 <- subset_metanalysis(dataset=list_array, adjpval = 0.05, abslog2FC = log2(1.5), max_n_genes = 500 )
#find co-DE genes for the DYRK1A gene (ensemblID="ENSG00000157540")
head(find_coDE(gene="ENSG00000157540",dataset=list_array.05_fc1.5_max500))
}
