% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_metaanalysis.R
\name{find_fold_changes_and_pvalues}
\alias{find_fold_changes_and_pvalues}
\title{Retrieve fold changes and adjusted-pvalues}
\usage{
find_fold_changes_and_pvalues(gene, dataset)
}
\arguments{
\item{gene}{a gene (character) included in one of the sublists in \code{dataset}.}

\item{dataset}{list of 3 lists, the first with gene IDs (called \code{names}), the second with the adjusted p-values for each genes (called \code{adjpval}) and the third with log2 fold changes (called \code{log2FC}). Each one of these three lists have a number of sublists corresponding to the gene IDs, adjusted p-values, and log2FCs, respectively from the differential expression analyses of each comparison.}
}
\value{
a dataframe with 2 columns: \code{log2FC} and \code{adjpval}. Each row is a different comparison where the given gene is found DE.
}
\description{
Retrieve for a given genes all the log2FCs and adjusted-p-values across the different comparison(s)
}
\examples{
# create a list of lists were only the first 500 most significant genes with adjusted p-value < 0.05 and fold change >1.5 or < -1.5 are included
data(list_array) #load data
list_array.05_fc1.5_max500 <- subset_metanalysis(dataset=list_array, adjpval = 0.05, abslog2FC = log2(1.5), max_n_genes = 500 )
#retrieve log2FCs and adjusted-p-values for the DYRK1A gene (ensemblID="ENSG00000157540")
find_fold_changes_and_pvalues(gene="ENSG00000157540",dataset=list_array.05_fc1.5_max500)
}
