% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_metaanalysis.R
\name{find_gene_statistics}
\alias{find_gene_statistics}
\title{Retrieve statistics for all or a subset of genes present in the meta-analysis}
\usage{
find_gene_statistics(dataset, genes = NULL, top = NULL)
}
\arguments{
\item{dataset}{list of 3 lists, the first with gene IDs (called \code{names}), the second with the adjusted p-values for each genes (called \code{adjpval}) and the third with log2 fold changes (called \code{log2FC}). Each one of these three lists have a number of sublists corresponding to the gene IDs, adjusted p-values, and log2FCs, respectively from the differential expression analyses of each comparison.}

\item{genes}{vector of gene names. Limit the results only to the elements of \code{genes}. Cannot be set together with \code{top}.}

\item{top}{integer. Limit the results to the top number of genes with the highest occurence in the dataset. Cannot be set together with \code{genes}.}
}
\value{
a dataframe with 6 columns and as many rows as genes. The column \code{names} reports the gene names used in the dataset. \code{occurences} indicates the number of times a gene is found differentially expressed. \code{median_log2FC}, \code{mean_log2FC} and \code{sd_log2FC}: median, mean, and standard deviations of all the log2 fold changes in which the gene is encountered across the comparisons. \code{pseudotscore}: this statistics is calculated as \code{mean_log2FC/(sd_log2FC/sqrt(occurences))}. Genes with a high pseudo-t-score have high absolute fold change that does not vary across comparisons and are found differentially expressed in a relatively high number of comparisons.
}
\description{
For each gene retrieve mean, median, and standard deviation of the log2FC and the number of times it is ecountered differentially expressed in the various comparisons.
}
\examples{
# create a list of lists were only the first 500 most significant genes with adjusted p-value < 0.05 and fold change >1.5 or < -1.5 are included
data(list_array) #load data
list_array.05_fc1.5_max500 <- subset_metanalysis(dataset=list_array, adjpval = 0.05, abslog2FC = log2(1.5), max_n_genes = 500 )
#retrieve the statistics for the most frequent 500 DE genes
find_gene_statistics(dataset=list_array.05_fc1.5_max500, top=500)
}
