% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipsc_deseq2}
\alias{ipsc_deseq2}
\title{Example of a differential expression analysis generated with \code{DESeq2}.}
\format{
A data frame with 58336 rows and 6 variables:
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE52249}
}
\usage{
ipsc_deseq2
}
\description{
mRNA-seq profiling of iPS cells (4 euploid and 3 trisomy 21) derived from fibroblasts of monozygotic twins discordant for trisomy 21
}
\keyword{datasets}
