% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{list_array}
\alias{list_array}
\title{Example of a dataset generated from 124 different differential expression analyses}
\format{
list of 3 lists called \code{names},  \code{adjpval}, and \code{log2FC} containing the same sublistis each with the same number of ordered elements:
\itemize{
\item \emph{names} list of 124 lists of gene names, each corresponding to a differnt comparison
\item \emph{adjpval} list of 124 lists of adjusted p-values, each corresponding to a differnt comparison
\item \emph{log2FC} list of 124 lists of log2 fold changes, each corresponding to a differnt comparison
}
}
\usage{
list_array
}
\description{
The dataset include microarray and RNA-seq experiments comparing trisomic samples versus euploid samples coming from differnt tissue and organs both in Homo sapines and Mus musculus, with the final goal to detect the genes most consistently differentially expressed in Down syndrome. Only genes with an adjusted p-value <0.1 are reported.
}
\keyword{datasets}
