% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_metaanalysis.R
\name{plot_gene_coDE_newtork}
\alias{plot_gene_coDE_newtork}
\title{Plot a network of co-DE genes}
\usage{
plot_gene_coDE_newtork(
  dataset = NULL,
  genes = NULL,
  top = NULL,
  code_matrix = NULL,
  gene_stats_matrix = NULL,
  node_color = c("mean", "median", "pseudotscore"),
  node.names = TRUE
)
}
\arguments{
\item{dataset}{list of 3 lists, the first with gene IDs (called \code{names}), the second with the adjusted p-values for each genes (called \code{adjpval}) and the third with log2 fold changes (called \code{log2FC}). Each one of these three lists have a number of sublists corresponding to the gene IDs, adjusted p-values, and log2FCs, respectively from the differential expression analyses of each comparison.}

\item{genes}{it is taken into consideration only when \code{dataset} is not \code{NULL}. Limit the search only to the desired genes. Cannot be set together with \code{top}.}

\item{top}{it is taken into consideration only when \code{dataset} is not \code{NULL}. Integer that limits the results to the top number of genes with the highest frequence in the dataset. Cannot be set together with \code{genes}.}

\item{code_matrix}{dataframe of co-DE events only used when \code{dataset=NULL} and together with \code{gene_stats_matrix}. Can be produced using the function \code{find_coDE}. The gene should match the one present in \code{gene_stats_matrix}.}

\item{gene_stats_matrix}{dataframe of statistics only used when \code{dataset=NULL} and together with \code{code_matrix}. Can be produced using the function \code{find_gene_statistics}.}

\item{node_color}{color nodes based on the genes' mean (\code{"mean"}) or median (\code{"median"}) log2 fold changes or based on their pseudo-t-scores (\code{"pseudotscore"}).}

\item{node.names}{whether to print (\verb{=TRUE}) or not (\verb{=FALSE}) node names.}
}
\value{
plot a network where each node is a gene (node size corresponds to the number of comparisons in which a given gene is found co-DE) and each edge represent an event of co-differential expression (edge thickeness corresponds to the number of comparison where the genes are found co-DE). Nodes are colored based on log2FC (median or mean across comparisons) or pseudo-t-score (for more information check documentation for \code{find_gene_statistics}). Also returns the co-DE graph object produced by the igraph package.
}
\description{
Plot a network where each node is a gene and each edge represent an event of co-differential expression.
}
\examples{
# create a list of lists were only the first 500 most significant genes with adjusted p-value < 0.05 and fold change >1.5 or < -1.5 are included
data(list_array) #load data
list_array.05_fc1.5_max500 <- subset_metanalysis(dataset=list_array, adjpval = 0.05, abslog2FC = log2(1.5), max_n_genes = 500)
#plot the co-DE network for the most frequent 10 genes and color the nodes based on their mean log2FC
plot_gene_coDE_newtork(dataset=list_array.05_fc1.5_max500, top = 10, node_color="mean")
}
