% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_metaanalysis.R
\name{subset_metanalysis}
\alias{subset_metanalysis}
\title{Subsetting the dataset}
\usage{
subset_metanalysis(dataset, adjpval = 0.1, max_n_genes = Inf, abslog2FC = 0)
}
\arguments{
\item{dataset}{list of 3 lists, the first with gene IDs (called \code{names}), the second with the adjusted p-values for each genes (called \code{adjpval}) and the third with log2 fold changes (called \code{log2FC}). Each one of these three lists have the number of sublists corresponding to the results of the differential expression analyses for each contrasts corresponding respectively to the geneIDs, adjusted p-values, and log2FCs.}

\item{adjpval}{desired adjusted p-value threshold for subsetting the data}

\item{max_n_genes}{maximum number of genes to be kept. It is used only if for a given contrasts there are more than \code{max_n_genes} genes that pass the chosen filtered. Ranking is based on the adjusted p-values and ties are included so there might be more than \code{max_n_genes} included.}

\item{abslog2FC}{desired absolute log2 fold change threshold for subsetting the data}
}
\value{
A list of lists structured identically to \code{dataset} but where the genes that do not pass the filters are excluded.
}
\description{
Subset DE analysis data for a bunch of comparisons based on fold change and p-values
}
\examples{
# create a list of lists were only the first 500 most significant genes with adjusted p-value < 0.05 and fold change >1.5 or < -1.5 are included
data(list_array) #load data
list_array.05_fc1.5_max500 <- subset_metanalysis(dataset=list_array, adjpval = 0.05, abslog2FC = log2(1.5), max_n_genes = 500 )
}
