% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{thymus_limma}
\alias{thymus_limma}
\title{Example of a differential expression analysis generated with \code{limma}.}
\format{
A data frame with 20994 rows and 16 variables:
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE69210}
}
\usage{
thymus_limma
}
\description{
Thymus transcriptomic profiles of patients with and without Down syndrome were compared in order to identify differentially expressed transcripts.
}
\keyword{datasets}
